Fix compiler warning
warning: passing argument 1 of 'ldv_is_err' makes pointer from integer without a cast [-Wint-conversion]

+++ presets/jobs/specifications/linux/err.c
@@ -53,5 +53,5 @@ long ldv_ptr_err(const void *ptr)
 
 bool ldv_is_err_or_null(const void *ptr)
 {
-	return !ptr || ldv_is_err((unsigned long)ptr);
+	return !ptr || ldv_is_err(ptr);
 }
